+++ CHANGELOG.md
@@ -9,11 +9,12 @@
 - Made some visual changes to `/serverinfo`
 - Moved role related commands to new Roles cog.
 - `/roles` now sorts results by position and shows members with the role.
+- New cog: `Settings`, to change all server specific settings: `/settings`
+    - `/setworld` moved to `/settings world`
 - Many changes to command names and aliases:
     - `/item`: `checkprice` alias removed, `items` alias added.
     - `/monster`: `mon` alias removed.
     - `/spell`: `spells` alias added.
-    - `/setworld`: `trackworld` alias added.
     - `/seteventchannel`: `seteventschannel` alias removed.
     - `/setleveldeathschannel`: `setlevelchannel`, `setdeathchannel`, `setleveldeathchannel` aliases removed.
     - `/server`: `server_info` alias removed.